Update README and HopSkipJump synthesizer
Adds functioning dummy synthesizer 'hop'

Turn dummy 'hop' into new 'hopskipjump" synthesizer

Directly link to issues; Reword design objectives
