+++ examples/extract_mnist_gpu.py
@@ -32,8 +32,7 @@ attack = ModelExtractionAttack(
     (1, 28, 28, 1),
     10,
     (1, 3, 28, 28),
-    "hop",
-    # "copycat",
+    "copycat",
     ImagenetTransferLearning,
     1000,
     emnist_train,
