+++ src/privacyraven/extraction/synthesis.py
@@ -1,10 +1,9 @@
 import torch
 from art.attacks.evasion import BoundaryAttack, HopSkipJump
-from art.estimators.classification import BlackBoxClassifier
 from sklearn.model_selection import train_test_split
 from tqdm import tqdm
 
-from privacyraven.utils.model_creation import NewDataset
+from privacyraven.utils.model_creation import NewDataset, set_evasion_model
 from privacyraven.utils.query import reshape_input
 
 synths = dict()
@@ -27,6 +26,7 @@ def synthesize(func_name, seed_data_train, seed_data_test, *args, **kwargs):
     Returns:
         Three NewDatasets containing synthetic data"""
     func = synths[func_name]
+    print("Time to synthesize")
     x_train, y_train = func(seed_data_train, *args, **kwargs)
     x_test, y_test = func(seed_data_test, *args, **kwargs)
     print("Synthesis complete")
@@ -39,28 +39,6 @@ def synthesize(func_name, seed_data_train, seed_data_test, *args, **kwargs):
     synth_valid = NewDataset(x_valid, y_valid)
     synth_test = NewDataset(x_test, y_test)
     return synth_train, synth_valid, synth_test
-
-
-def set_evasion_model(query, victim_input_shape, victim_input_targets):
-    """Defines the threat model for an evasion attack"""
-    config = BlackBoxClassifier(
-        predict=query,
-        input_shape=victim_input_shape,
-        nb_classes=victim_input_targets,
-        clip_values=(0, 255),
-        preprocessing_defences=None,
-        postprocessing_defences=None,
-        preprocessing=None,
-    )
-    return config
-
-
-def init_hopskipjump(config, data, limit=50):
-    """Runs the HopSkipJump evasion attack
-
-    Arxiv Paper: https://arxiv.org/abs/1904.02144"""
-    attack = HopSkipJump(config, False, max_iter=limit, max_eval=100, init_eval=10)
-    return attack.generate(data)
 
 
 def get_data_limit(data):
@@ -96,7 +74,7 @@ def copycat(
     else:
         limit = data_limit
 
-    print(limit)
+    # print(limit)
 
     for i in tqdm(range(0, limit)):
         if i == 0:
@@ -120,10 +98,9 @@ def copycat(
             y = torch.cat((y, yi))
 
     # print(f"Dataset Created: {x.shape}; {y.shape}")
+    print(x.dtype)
+    print(y.dtype)
     return x, y
-
-
-# def copycat_wrap(
 
 
 @register_synth
@@ -135,37 +112,33 @@ def hopskipjump(
     substitute_input_shape,
     victim_input_targets,
 ):
-    """Generates a dataset from unlabeled data using the hopskipjump attack"""
-    x, y = copycat(
+    """Runs the HopSkipJump evasion attack
+
+    Arxiv Paper: https://arxiv.org/abs/1904.02144"""
+    config = set_evasion_model(query, victim_input_shape, victim_input_targets)
+    internal_limit = int(query_limit * 0.5)
+    evasion_limit = int(query_limit * 0.5)
+    attack = HopSkipJump(
+        config,
+        False,
+        norm="inf",
+        max_iter=evasion_limit,
+        max_eval=evasion_limit,
+        init_eval=10,
+    )
+    X, y = copycat(
         data,
         query,
-        query_limit,
+        internal_limit,
         victim_input_shape,
         substitute_input_shape,
         victim_input_targets,
     )
-    x = x.to(torch.float32)
-    import pdb
-
-    pdb.set_trace()
-    config = set_evasion_model(query, victim_input_shape, victim_input_targets)
-    x_adv, y_adv = init_hopskipjump(config, data)
-    x = torch.cat((x, x_adv))
-    y = torch.cat((y, y_adv))
-    return x, y
-
-
-@register_synth
-def seeded_hopskipjump(
-    data,
-    query,
-    query_limit,
-    victim_input_shape,
-    substitute_input_shape,
-    victim_input_targets,
-):
-    """Generates a dataset from labeled data using the hopskipjump attack"""
-    data = data.to(torch.float32)
-    config = set_evasion_model(query, victim_input_shape, victim_input_targets)
-    x_adv, y_adv = init_hopskipjump(config, data, query_limit)
-    return x_adv, y_adv
+    print(X.shape)
+    result = attack.generate(X)
+    result = torch.as_tensor(result)
+    result = result.clone().detach()
+    print(result.shape)
+    y = torch.Tensor([query(x) for x in result])
+    y = y.long()
+    return result, y
